rfg: Performance improvements from vim-fold-js
This commit replaces the foldexpr algorithm
with a simpler, less computationally expensive one.

A quick test on this sample spec from Zammad[0]
shows a 40% reduction in CPU time:

    # before
    $ time vim spec/models/user_spec.rb -c ":quit"

    real    0m2.150s
    user    0m1.793s
    sys     0m0.323s

    # after
    $ time vim spec/models/user_spec.rb -c ":quit"

    real    0m1.266s
    user    0m1.092s
    sys     0m0.150s

This improvement is achieved by leveraging
vimscript's built-in `searchpair()` method
to find matching `do`/`end` keywords.

(The foldexpr finds `end`s and looks for their matching `do`s.
Prior to this commit, we would start from an `end`
and recursively walk backwards through the file line-by-line
until we hit the `do` we were looking for.
Not only was this time-consuming,
it also often risked exceeding the 'maxfuncdepth' setting.)

=== NOTE: This commit breaks tests (false negatives).

[0]: https://github.com/zammad/zammad/blob/develop/spec/models/user_spec.rb